Move decorator utilities to the rbac folder
