+++ framework/wazuh/rbac/tests/data/utils/test_lists
@@ -1,0 +1,5 @@
+test-wazuh-w:write
+test-wazuh-r:read
+test-wazuh-a:attribute
+test-wazuh-x:execute
+test-wazuh-c:command
