+++ framework/wazuh/rbac/tests/data/utils/test_rules.xml
@@ -1,0 +1,17 @@
+<!-- Local rules -->
+
+<!-- Modify it at your will. -->
+
+<!-- Example -->
+<group name="local,">
+
+  <!--
+   NEW RULE 
+   -->
+   <rule id="100001111" level="5">
+     <if_sid>5716</if_sid>
+     <srcip>1.1.1.1</srcip>
+     <description>sshd: authentication failed from IP 1.1.1.1.</description>
+     <group>authentication_failed,pci_dss_10.2.4,pci_dss_10.2.5,</group>
+   </rule>
+</group>
